# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "description": "Successful auth response.",
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "user": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "User",
            "type": "object"
          }
        },
        "required": [
          "accessToken",
          "user"
        ],
        "title": "AuthResponse",
        "type": "object"
      },
...
      "GoogleLoginRequest": {
        "description": "Credential posted by Google Identity Services on the frontend.",
        "properties": {
          "credential": {
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          }
        },
        "required": [
          "credential"
        ],
        "title": "GoogleLoginRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Scope Stock Research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/google": {
      "post": {
        "description": "Verify Google identity and issue a Scope JWT plus revocable session cookie.",
        "operationId": "login_with_google_api_v1_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login With Google",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "description": "Return the current authenticated user.",
        "operationId": "get_me_api_v1_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "scope_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Me Api V1 Auth Me Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}